Changed arguments of VeoProc.read_mem and VeoProc.write_mem. Local (VH side) buffers now can be any Python object that supports the buffer protocol instead of only numpy ndarrays.
Added OnStack() class that allows easy passing of objects as
temporary elements allocated on the stack and accessed by reference.
Simply use OnStack(buff) as the proper argument. buff can be any
python object that supports the buffer protocol.

Added examples for testing memory transfer and passing of arguments
on stack, for example a struct built with cffi.